+++ mockspresso-dagger/src/main/java/com/episode6/hackit/mockspresso/dagger/DaggerLazyMaker.kt
@@ -3,18 +3,15 @@ package com.episode6.hackit.mockspresso.dagger
 import com.episode6.hackit.mockspresso.api.DependencyProvider
 import com.episode6.hackit.mockspresso.api.SpecialObjectMaker
 import com.episode6.hackit.mockspresso.reflect.DependencyKey
-import com.episode6.hackit.mockspresso.reflect.TypeToken
 import com.episode6.hackit.mockspresso.reflect.genericParameterKey
+import com.episode6.hackit.mockspresso.reflect.isParameterizedGeneric
 import dagger.Lazy
-import java.lang.reflect.ParameterizedType
 
 /**
  * An implementation of SpecialObjectMaker for [dagger.Lazy].
  */
 internal class DaggerLazyMaker : SpecialObjectMaker {
-  override fun canMakeObject(key: DependencyKey<*>): Boolean =
-      key.typeToken.rawType == Lazy::class.java &&
-          key.typeToken.type is ParameterizedType
+  override fun canMakeObject(key: DependencyKey<*>): Boolean = key.isParameterizedGeneric(Lazy::class)
 
   @Suppress("UNCHECKED_CAST")
   override fun <T> makeObject(dependencyProvider: DependencyProvider, key: DependencyKey<T>): T? {