+++ buildSrc/src/main/groovy/DeployableJavaProject.groovy
@@ -1,66 +1,48 @@
 import org.gradle.api.Project
-import org.gradle.api.artifacts.Configuration
-import org.gradle.api.artifacts.ExternalDependency
-import org.gradle.api.artifacts.ProjectDependency
 import org.gradle.api.plugins.PluginContainer
-import org.gradle.api.tasks.Sync
 
 class DeployableJavaProject extends JavaProject {
+
+  @Override
+  protected void applyDeployablePlugin(PluginContainer plugins) {
+    plugins.with {
+      apply 'org.jetbrains.dokka'
+      apply 'maven-publish'
+      apply 'signing'
+    }
+  }
 
   @Override
   void apply(Project project) {
     super.apply(project)
 
-    project.task(Tasks.SYNC_DOCS, type: Sync) {
-      from project.tasks.dokka.outputDirectory
-      into "${project.rootProject.rootDir}/docs/javadocs/${project.name}"
-      dependsOn project.tasks.dokka
+    project.with {
+      // dokkaHtml is used to generate javadoc jars for maven uploads
+      tasks.dokkaHtml {
+        dokkaSourceSets {
+          configureEach {
+            jdkVersion.set(8)
+            reportUndocumented.set(false)
+            noStdlibLink.set(true)
+            noJdkLink.set(true)
+          }
+        }
+      }
 
-      // this file doesn't get generated correctly in the current version of dokka
-      // each link in it has two nested <a href> tags, where the inner-one is correct
-      // but the outer one references the absolute path where the docs were first
-      // generated (i.e. my laptop).
-      // This file also isn't linked anywhere from the rest of the javadocs
-      // so excluding it for now should be NBD.
-      excludes += "${project.name}/index-outline.html"
-    }
-
-    project.tasks.dokka {
-      jdkVersion = 8
-    }
-
-    project.configurations.all { Configuration config ->
-      //noinspection UnstableApiUsage
-      config.withDependencies { dependencies ->
-
-        // Each module's javadocs are generated separately and can't see each other (atm).
-        // This block links to our already hosted docs from the previous release.
-        // It's not ideal, but its better than nothing.
-        dependencies.findAll { it instanceof ProjectDependency }
-            .forEach { dep ->
-              project.tasks.dokka {
-                externalDocumentationLink {
-                  url = new URL("https://episode6.github.io/mockspresso/javadocs/${dep.name}/${dep.name}/")
-                }
-              }
-            }
-
-        // Include links to javadocs of external dependencies when possible
-        dependencies.findAll { it instanceof ExternalDependency }
-            .collect { (ExternalDependency)it }
-            .collect { ExternalDocs.getUrlFor(project, it) }
-            .findAll { it != null }
-            .forEach { extUrl ->
-              project.tasks.dokka { externalDocumentationLink { url = extUrl } }
-            }
+      // dokkaHtmlPartial is used to generate a complete set of docs for the web
+      tasks.dokkaHtmlPartial {
+        dokkaSourceSets {
+          configureEach {
+            jdkVersion.set(8)
+            reportUndocumented.set(false)
+            noStdlibLink.set(false)
+            noJdkLink.set(false)
+          }
+        }
       }
     }
 
-
-  }
-
-  @Override
-  protected void applyDeployablePlugin(PluginContainer plugins) {
-    plugins.apply 'com.episode6.hackit.deployable.kt.jar'
+    ExternalDocs.configureSourcesFor(project)
+    MavenConfiguration.setup(project)
   }
 }