+++ buildSrc/src/main/groovy/ExternalDocs.groovy
@@ -1,47 +1,58 @@
-import org.gradle.api.GradleException
 import org.gradle.api.Project
+import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.ExternalDependency
 
 import javax.annotation.Nullable
 
 /**
- *
+ * Configurations/urls for external documentation
  */
 class ExternalDocs {
-  private static final String VERSION_TAG = "<version>"
-  private static final Map<String, String> DOCS_MAP = [
-      'junit:junit'                                 : "https://junit.org/junit4/javadoc/$VERSION_TAG/",
-      'com.google.dagger:dagger'                    : "https://google.github.io/dagger/api/$VERSION_TAG/",
-      'javax.inject:javax.inject'                   : "https://docs.oracle.com/javaee/7/api/",
-      'org.easymock:easymock'                       : "http://easymock.org/api/",
-      'com.google.guava:guava'                      : "https://google.github.io/guava/releases/$VERSION_TAG/api/docs/",
-      'org.mockito:mockito-core'                    : "https://static.javadoc.io/org.mockito/mockito-core/$VERSION_TAG/",
-      'org.powermock:powermock-module-junit4'       : "https://static.javadoc.io/org.powermock/powermock-module-junit4/$VERSION_TAG/",
-      'org.powermock:powermock-api-easymock'        : "https://static.javadoc.io/org.powermock/powermock-api-easymock/$VERSION_TAG/",
-      'org.powermock:powermock-module-junit4-rule'  : "https://static.javadoc.io/org.powermock/powermock-module-junit4-rule/$VERSION_TAG/",
-      'org.powermock:powermock-classloading-xstream': "https://static.javadoc.io/org.powermock/powermock-classloading-xstream/$VERSION_TAG/",
-      'org.powermock:powermock-api-mockito2'        : "https://static.javadoc.io/org.powermock/powermock-api-mockito2/$VERSION_TAG/",
-  ]
 
-  static @Nullable URL getUrlFor(Project project, ExternalDependency dependency) {
+  static String JAVAX_INJECT_DOCS_URL = "https://docs.oracle.com/javaee/7/api/"
+
+  static @Nullable String getUrlFor(ExternalDependency dependency) {
     String key = "${dependency.group}:${dependency.name}"
-    if (!DOCS_MAP.containsKey(key)) {
-      return null
+    String version = dependency.getVersion()
+    switch (key) {
+      case 'junit:junit':
+        return "https://junit.org/junit4/javadoc/$version/"
+      case 'javax.inject:javax.inject':
+        return JAVAX_INJECT_DOCS_URL
+      case 'org.easymock:easymock':
+        return "http://easymock.org/api/"
+      case 'com.google.guava:guava':
+        return "https://google.github.io/guava/releases/$version/api/docs/"
+      case 'org.mockito:mockito-core':
+        return "https://static.javadoc.io/org.mockito/mockito-core/$version/"
+      case 'org.powermock:powermock-module-junit4':
+        return "https://static.javadoc.io/org.powermock/powermock-module-junit4/$version/"
+      case 'org.powermock:powermock-api-easymock':
+        return "https://static.javadoc.io/org.powermock/powermock-api-easymock/$version/"
+      case 'org.powermock:powermock-module-junit4-rule':
+        return "https://static.javadoc.io/org.powermock/powermock-module-junit4-rule/$version/"
+      case 'org.powermock:powermock-classloading-xstream':
+        return "https://static.javadoc.io/org.powermock/powermock-classloading-xstream/$version/"
+      case 'org.powermock:powermock-api-mockito2':
+        return "https://static.javadoc.io/org.powermock/powermock-api-mockito2/$version/"
     }
-
-    String url = DOCS_MAP.get(key)
-    @Nullable String version = lookupVersion(project, key)
-    if (version != null) {
-      url = url.replace(VERSION_TAG, version)
-    }
-    return new URL(url)
+    return null;
   }
 
-  private static @Nullable String lookupVersion(Project project, String key) {
-    try {
-      return project.gdmcVersion(key)
-    } catch (GradleException ignore) {
-      return null
+  static void addLink(Project project, String externalUrl) {
+    project.tasks.dokkaHtmlPartial {
+      dokkaSourceSets { configureEach { externalDocumentationLink { url.set(new URL(externalUrl)) } } }
+    }
+  }
+
+  static void configureSourcesFor(Project project) {
+    project.configurations.all { Configuration config ->
+      config.withDependencies { dependencies ->
+        dependencies.withType(ExternalDependency)
+            .collect { getUrlFor(it) }
+            .findAll { it != null }
+            .forEach { extUrl -> addLink(project, extUrl) }
+      }
     }
   }
 }