+++ buildSrc/src/main/groovy/JavaProject.groovy
@@ -1,4 +1,3 @@
-import org.gradle.api.JavaVersion
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.plugins.PluginContainer
@@ -12,14 +11,13 @@ class JavaProject implements Plugin<Project> {
         apply 'java-library'
         apply 'kotlin'
         applyDeployablePlugin(it)
-        apply 'com.episode6.hackit.gdmc'
       }
 
       sourceCompatibility = 1.8
       targetCompatibility = 1.8
 
       dependencies {
-        implementation 'org.jetbrains.kotlin:kotlin-stdlib'
+        implementation libs.kotlin.stdlib
       }
     }
   }