+++ buildSrc/src/main/groovy/Tasks.groovy
@@ -1,6 +1,0 @@
-/**
- *
- */
-class Tasks {
-  public static final String SYNC_DOCS = "syncDocs"
-}